article-debate-system/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
langchain
langgraph
langchain_groq
pydantic
python-dotenv
requests